Make FindUpsPackage a wrapper around find_package and pkg-config
The logic here may not be quite complete. FindUpsPackage seems to
do basically three things:
(1) check if the requested package has been set up with at least
 the requested version, if given. (The qualifier isn't checked, but
 the flavor-qualifier is.)
(2) Load the package's <Package>Config.cmake file. This, in turn,
 does quite a lot:
 (a) sets product-related variables (library names etc.)
 (b) finds dependent packages via find_ups_product (so it is usually
   called recursively)
 (c) Loads libraries via cet_find_library
 (d) Adds the project's include director(y|ies) via include_directories
(3) Add the UPS product's include directory (<PRODUCT>_INC) to the
 list of include locations.
It looks like (3) might be redundant. If the Config.cmake file is
created correctly, the product's include path(s) have already been added
in (2).
Our new version simplifies (1), still does (2) exactly as before, and
skips (3) (unless locating the package via pkg-config).
Things may not be set up fully yet for the pkg-config case. We may have
to define additional variables, like <PRODUCT>_DIR etc.

And a bit of a hack: Since the sqlite UPS package is called "sqlite",
explicitly check for this name and replace it with "sqlite3", because
that's what pkg-config needs.