buildtools: Don't request deprecated boost::signals (removed in Boost 1.69)
boost::signals is really old and has long been replaced with boost::signals2.
Just removing it as a required component doesn't seem to do any harm;
there are no link failures with any of the art suite components when I do it.
So this looks like cruft. Wouldn't be surprised if there's more like it.