art: Fix failing FPCTest_Invalid test
The code attempts to generate an FPE_Invalid exception by calling
std::log(-1.0). With newer compilers and C++ standards, it seems
like this expression is already evaluated at compile time. To
prevent this optimization, we move it into a separately compiled
function, arttest::logit(x), similar to what is is done for some
of the other FP exception tests.