Support reading environment variables in workbook modulefile
If ARIEL_WORKDIR is set, it is used as the build directory. If
non-existent, it is created. "bin" and "lib" subdirectories
inside the build directory are also created if necessary.

The output directory for data files is set according to the
following logic (see also examples/workbook/admin/
check_outputDir_symlink.in):

- If ART_WORKBOOK_OUTPUT is set, it is used as the output directory.
- Otherwise, if ART_WORKBOOK_OUTPUT_BASE is set, then the output
  goes into $ART_WORKBOOK_OUTPUT_BASE/$USER/art_workbook_output.
- Otherwise, <build-dir>/../output is used for output.
- If the output directory does not exist, it will be created
  at CMake time.

For the moment, we retain the historical environment variable
naming "CET*" and "ART*". In the future, these will change to
"ARIEL*".

Also, this commit fixes a bug in the previous implementation where
ART_WORKBOOK_OUTPUT was not unset when unloading the module.

Note:
% module (display|show) workbook
may print incorrect results. There seems to be a bug in the modules
package where in display mode [getenv VAR] always returns "$VAR"
instead of the value of $VAR or "_UNDEFINED_", as it does in all
other modes.